Add Explorer of inner classes
Inner classes are funny. You can nest them as deep as you want. You can
go full "Inception" on them: "InnerCeption" (sorry, I had to make this
bad pun).

But seriously: the syntax of creating an inner class is funny because it
requires an already created outer object. And instead of calling the
constructor with new Outer().Inner() (as a normal method) we need to
cram "new" in there to get this: new Outer().new Inner();